requirements.txt
bs4
scrapy
google-cloud
google-cloud-bigquery
fastapi